TAP-103: Global styles paired down
