--- packages/article/styles/article-global-style.js
@@ -1,29 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const globalStyle = StyleSheet.create({
-  articleMainContentRow: {
-    paddingLeft: 10,
-    paddingRight: 10
-  },
-  articleMiddleContainer: {
-    paddingTop: 9
-  },
-  articleLabel: {
-    paddingTop: 4,
-    paddingBottom: 3
-  },
-  articleFlag: {
-    marginBottom: 11,
-    flexDirection: "row"
-  },
-  articleFlagContainer: {
-    marginRight: 14
-  },
-  articleMeta: {
-    marginBottom: 20,
-    borderBottomColor: "#d0cece",
-    borderBottomWidth: StyleSheet.hairlineWidth
-  }
-});
-
-export default globalStyle;
