+++ packages/article/styles/body/shared.js
@@ -8,6 +8,10 @@ const sharedStyles = {
     fontSize: 17,
     marginBottom: 25,
     color: "#333333"
+  },
+  articleMainContentRow: { // article.js
+    paddingLeft: 10,
+    paddingRight: 10
   }
 };
