+++ packages/article/styles/meta/shared.js
@@ -1,10 +1,18 @@
 import { StyleSheet } from "react-native";
 
 const sharedStyles = {
+  articleMeta: {
+    marginBottom: 20,
+    borderBottomColor: "#d0cece",
+    borderBottomWidth: StyleSheet.hairlineWidth
+  },
+  articleMiddleContainer: { // meta
+    paddingTop: 9
+  },
   articleMetaElement: {
     borderTopColor: "#d0cece",
     borderTopWidth: StyleSheet.hairlineWidth
-  }
+  },
 };
 
 export default sharedStyles;
