+++ packages/article/styles/shared.js
@@ -1,0 +1,10 @@
+import { StyleSheet } from "react-native";
+
+const globalStyle = StyleSheet.create({
+  articleMainContentRow: {
+    paddingLeft: 10,
+    paddingRight: 10
+  },
+});
+
+export default globalStyle;
